Check if file exist in backwards.php

+++ PodcastGenerator/core/backwards.php
@@ -32,15 +32,24 @@ function backwards_2_7_to_3_0($absoluteurl) {
     ];
     // Delete files
     for($i = 0; $i < sizeof($filesToDelete); $i++) {
-        unlink($absoluteurl . $filesToDelete[$i]);
+        if(file_exists($absoluteurl . $filesToDelete[$i])) {
+            unlink($absoluteurl . $filesToDelete[$i]);
+        }
     }
     // Delete directories
     for($i = 0; $i < sizeof($dirsToDelete); $i++) {
         array_map('unlink', glob($absoluteurl . $dirsToDelete[$i]."/*.*"));
         rmdir($absoluteurl . $dirsToDelete[$i]);
     }
+    // Remove tabs in the config
+    $c = file_get_contents('config.php');
+    $c = str_replace("\t", '', $c);
+    file_put_contents('config.php', $c);
     // Update theme
     updateConfig('config.php', 'theme_path', 'themes/default/');
     // Update version
     updateConfig($absoluteurl . 'config.php', 'podcastgen_version', $version);
+    sleep(0.5);
+    header('Location: index.php');
+    die();
 }
